+++ packages/navigation/src/components/js/GlobalItem.jsx
@@ -2,7 +2,7 @@
 import React, { PureComponent } from 'react';
 import GlobalItemInner from '../styled/GlobalItemInner';
 import DefaultLinkComponent from './DefaultLinkComponent';
-import type { ReactElement } from '../../types';
+import type { ReactElement, ReactClass } from '../../types';
 
 type Props = {|
   /** Element to be rendered inside the item. Should be an atlaskit icon. */
@@ -13,7 +13,7 @@ type Props = {|
   size?: 'small' | 'medium' | 'large',
   /** Component to be used to create the link in the global item. A default
   component is used if none is provided. */
-  linkComponent?: () => ReactElement,
+  linkComponent?: ReactClass,
 |}
 
 export default class GlobalItem extends PureComponent {