+++ packages/navigation/stories/components/nested-navigation/BasicNestedNavigation.jsx
@@ -5,6 +5,7 @@ import ArrowLeftIcon from '@atlaskit/icon/glyph/arrow-left';
 import DashboardIcon from '@atlaskit/icon/glyph/dashboard';
 import SettingsIcon from '@atlaskit/icon/glyph/settings';
 import TrayIcon from '@atlaskit/icon/glyph/tray';
+import InlineDialog from '@atlaskit/inline-dialog';
 import Tooltip from '@atlaskit/tooltip';
 
 import {
@@ -73,27 +74,46 @@ export default class BasicNestedNavigation extends PureComponent {
 
   state = {
     stack: [BasicNestedNavigation.menu],
+    isHeaderInlineDialogOpen: false,
   }
 
-  getContainerHeaderComponent = () => (
-    <div>
-      <AkContainerTitle
-        href="#foo"
-        icon={
-          <img alt="nucleus" src={nucleusLogo} />
-        }
-        text="AtlasKit"
-        subText="Is the king"
+  getContainerHeaderComponent = () => {
+    const backButton = this.state.stack.length > 1 ? (
+      <AkNavigationItem
+        icon={<ArrowLeftIcon label="Back" />}
+        onClick={() => this.stackPop()}
+        text="Back" key="2"
       />
-      {this.state.stack.length > 1 ? (
-        <AkNavigationItem
-          icon={<ArrowLeftIcon label="Back" />}
-          onClick={() => this.stackPop()}
-          text="Back"
-        />
-      ) : null}
-    </div>
-  )
+    ) : null;
+
+    /* eslint-disable jsx-a11y/no-static-element-interactions */
+    return [
+      <InlineDialog
+        content={<div style={{ maxWidth: '200px' }}>Menu or something like the HipChat status form would go here</div>}
+        isOpen={this.state.isHeaderInlineDialogOpen}
+        position="bottom left"
+      >
+        <div
+          onClick={() => {
+            this.setState({ isHeaderInlineDialogOpen: !this.state.isHeaderInlineDialogOpen });
+          }}
+        >
+          <Tooltip key="1" position="right" description="Header tooltip text">
+            <AkContainerTitle
+              href="#foo"
+              icon={
+                <img alt="nucleus" src={nucleusLogo} />
+              }
+              text="AtlasKit"
+              subText="Is the king"
+            />
+          </Tooltip>
+        </div>
+      </InlineDialog>,
+      backButton,
+    ];
+    /* eslint-enable jsx-a11y/no-static-element-interactions */
+  };
 
   stackPush = (newPage) => {
     const stack = [...this.state.stack, newPage];