+++ packages/tag/src/components/Tag.jsx
@@ -1,4 +1,3 @@
-import PropTypes from 'prop-types';
 import React, { PureComponent } from 'react';
 
 import Chrome from './Chrome';
@@ -7,36 +6,68 @@ import Remove from './Remove';
 
 import Before from '../styled/Before';
 import Container from '../styled/Container';
+import type { appearanceType, tagColor, ReactElement } from '../types';
+
+const colorList = [
+  'standard',
+  'green',
+  'blue',
+  'red',
+  'purple',
+  'grey',
+  'teal',
+  'yellow',
+  'greenLight',
+  'blueLight',
+  'redLight',
+  'purpleLight',
+  'greyLight',
+  'tealLight',
+  'yellowLight',
+];
+
+type Props = {|
+  /** Set whether tags should be rounded. */
+  appearance?: appearanceType,
+  /** The color theme to apply, setting both background and text color. */
+  color?: tagColor,
+  /** Component to be rendered before the Tag. */
+  elemBefore?: ReactElement,
+  /** Text to be displayed in the tag. */
+  text: string,
+  /** uri or path. If provided, the tag will be a link.  */
+  href?: string,
+  /** Text display as the aria-label for remove text. Setting this makes the
+  tag removable. */
+  removeButtonText?: string,
+  /** Handler to be called before the tag is removed. If it does not return a
+  truthy value, the tag will not be removed. */
+  onBeforeRemoveAction?: () => null,
+  /** Handler to be called after tag is removed. Called with the string 'Post
+  Removal Hook'. */
+  onAfterRemoveAction?: () => null,
+|}
+
+type State = {
+  isRemoving: bool,
+  isRemoved: bool,
+  markedForRemoval: bool,
+  isFocused: bool,
+}
 
 export default class Tag extends PureComponent {
-  static propTypes = {
-    /** Set whether tags should be rounded. */
-    appearance: PropTypes.oneOf(['default', 'rounded']),
-    /** Component to be rendered before the Tag. */
-    elemBefore: PropTypes.node,
-    /** Text to be displayed in the tag. */
-    text: PropTypes.string.isRequired,
-    /** uri or path. If provided, the tag will be a link.  */
-    href: PropTypes.string,
-    /** Text display as the aria-label for remove text. Setting this makes the
-    tag removable. */
-    removeButtonText: PropTypes.string,
-    /** Handler to be called before the tag is removed. If it does not return a
-    truthy value, the tag will not be removed. */
-    onBeforeRemoveAction: PropTypes.func,
-    /** Handler to be called after tag is removed. Called with the string 'Post
-    Removal Hook'. */
-    onAfterRemoveAction: PropTypes.func,
-  }
+  props: Props // eslint-disable-line react/sort-comp
+  state: State // eslint-disable-line react/sort-comp
 
   static defaultProps = {
+    color: 'standard',
     appearance: 'default',
     elemBefore: null,
     onAfterRemoveAction: () => {},
     onBeforeRemoveAction: () => true,
   }
 
-  constructor(props) {
+  constructor(props: Props) {
     super(props);
     this.state = {
       isRemoving: false,
@@ -64,11 +95,21 @@ export default class Tag extends PureComponent {
 
   render() {
     const { isFocused, isRemoved, isRemoving, markedForRemoval } = this.state;
-    const { appearance, elemBefore, href, removeButtonText, text } = this.props;
+    const { appearance, elemBefore, href, removeButtonText, text, color } = this.props;
+
+    const safeColor = colorList.includes(color) ? color : 'standard';
 
     const isRemovable = Boolean(removeButtonText);
     const isRounded = appearance === 'rounded';
-    const styled = { isFocused, isRemovable, isRemoved, isRemoving, isRounded, markedForRemoval };
+    const styled = {
+      isFocused,
+      isRemovable,
+      isRemoved,
+      isRemoving,
+      isRounded,
+      markedForRemoval,
+      color: safeColor,
+    };
     const onAnimationEnd = e => isRemoving && this.handleRemoveComplete(e);
 
     return (